requirements.txt
streamlit
pandas
plotly
pyyaml
numpy
matplotlib